+++ src/components/Layout/Layout.test.tsx
@@ -1,11 +1,12 @@
-import { render, screen } from "@testing-library/react";
+import { screen } from "@testing-library/react";
 import React from "react";
+import { feedbackRender } from "~/src/__tests__/testUtils/render";
 import Layout from "./Layout";
 jest.mock("next/router", () => require("next-router-mock"));
 
 describe("Layout component", () => {
   beforeEach(() => {
-    render(
+    feedbackRender(
       <Layout>
         <div>Text in layout body</div>
       </Layout>