+++ src/pages/_app.tsx
@@ -10,6 +10,7 @@ import { documentTitles } from "../constants/labels";
 import "@nypl/web-reader/dist/index.css";
 import { FeatureFlagProvider } from "../context/FeatureFlagContext";
 import { trackPageview } from "../lib/Analytics";
+import { FeedbackProvider } from "../context/FeedbackContext";
 
 /**
  * Determines if we are running on server or in the client.
@@ -81,7 +82,9 @@ function MyApp({ Component, pageProps }: AppProps) {
         <script src={appConfig.analytics} async />
       </Head>
       <FeatureFlagProvider>
-        <Component {...pageProps} />
+        <FeedbackProvider>
+          <Component {...pageProps} />
+        </FeedbackProvider>
       </FeatureFlagProvider>
     </>
   );